requirements.txt
python-telegram-bot[job-queue]
yt-dlp
Flask